typo fix and member add to dictionary uncommented
Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch yeaidk
Your branch is up to date with 'public/yeaidk'.

No commands done.
Next command to do (1 remaining command):
   pick dd8ba3b see the commit message below # empty
You are currently editing a commit while rebasing branch 'Leveling' on '89b2901'.

Changes to be committed:
	modified:   cogs/member_function.py
	modified:   cogs/reaction_roles.py
	modified:   commands/user_card.py

Changes not staged for commit:
	modified:   Logs/infos.log
	modified:   settings.py
	modified:   users.json